% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_functions.R
\name{analyse_zhang}
\alias{analyse_zhang}
\title{Internal Function: Zhang and Gant Analysis}
\usage{
analyse_zhang(dataref, dataquery, nref = NULL, nquery = NULL,
  ord.query = TRUE, permute = FALSE, B = 1e+05, ntop.pvalues = 20,
  ntop.scores = 20, basefilename = "analyseZhang", colour.query = NULL,
  legend.names = NULL, legend.cols = unique(colour.query),
  legend.pos = "topright", result.available = NULL, plot.type = "pdf",
  print.top = TRUE, labels = TRUE, which = c(1))
}
\arguments{
\item{dataref}{NA}

\item{dataquery}{NA}

\item{nref}{NA}

\item{nquery}{NA}

\item{ord.query}{NA}

\item{permute}{NA}

\item{B}{NA}

\item{ntop.pvalues}{NA}

\item{ntop.scores}{NA}

\item{basefilename}{NA}

\item{colour.query}{NA}

\item{legend.names}{NA}

\item{legend.cols}{NA}

\item{legend.pos}{NA}

\item{result.available}{NA}

\item{plot.type}{NA}

\item{print.top}{NA}

\item{which}{NA}
}
\value{
NA
}
\description{
\strong{Internal function not meant for end-users.} Internal function involved in the computation of the Zhang and Gant score when using multiple cores with \code{parallel}.
}
\keyword{internal}
