% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/method_structure.R
\docType{methods}
\name{CSanalysis,matrix,matrix,character-method}
\alias{CSanalysis,matrix,matrix,character-method}
\title{Connectivity Score Analysis.}
\usage{
\S4method{CSanalysis}{matrix,matrix,character}(refMat, querMat, type, ...)
}
\arguments{
\item{refMat}{Reference matrix}

\item{querMat}{Query matrix}

\item{type}{Type of Factor Analysis or Zhang & Gant ( \code{"CSfabia"}, \code{"CSmfa"}, \code{"CSpca"}, \code{"CSsmfa"} or \code{"CSzhang"})}

\item{...}{Additional parameters for analysis}
}
\value{
An object of the S4 Class \code{CSresult}.
}
\description{
Doing a CS analysis, interactively generating graphs. See specific type for additional parameteres.
}
\examples{
\dontrun{
data("dataSIM",package="CSFA")
Mat1 <- dataSIM[,c(1:6)]
Mat2 <- dataSIM[,-c(1:6)]

MFA_analysis <- CSanalysis(Mat1,Mat2,"CSmfa")
FABIA_analysis <- CSanalysis(Mat1,Mat2,"CSfabia")
ZHANG_analysis <- CSanalysis(Mat1,Mat2,"CSzhang")
}
}

