% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/method_structure.R
\docType{methods}
\name{CSanalysis,matrix,matrix,CSmfa-method}
\alias{CSanalysis,matrix,matrix,CSmfa-method}
\title{"CSmfa"}
\usage{
\S4method{CSanalysis}{matrix,matrix,CSmfa}(refMat, querMat, type = "CSmfa",
  ncp = 5, weight.col.mfa = NULL, row.w = NULL, which = c(1, 2, 3, 4,
  5), factor.plot = NULL, column.interest = NULL, color.columns = NULL,
  row.interest = NULL, gene.thresP = 1, gene.thresN = -1,
  thresP.col = "blue", thresN.col = "red", legend.names = NULL,
  legend.cols = NULL, result.available = NULL,
  basefilename = "analyseMFA", plot.type = "device")
}
\arguments{
\item{refMat}{Reference matrix (Rows = genes and columns = compounds)}

\item{querMat}{Query matrix}

\item{type}{\code{"CSmfa"}}

\item{ncp}{\emph{MFA Parameter:} Number of dimensions kept in the results (by default 5).}

\item{weight.col.mfa}{\emph{MFA Parameter:} Vector of weights, useful for HMFA method (by default, \code{NULL} and an MFA is performed).}

\item{row.w}{\emph{MFA Parameter:} An optional row weights (by default, a vector of 1 for uniform row weights).}

\item{which}{Choose one or more plots to draw. 1: Percentage of variance explained by factors, 2: Loadings for reference compounds, 3: Factor 1 VS Factor 2 : Loadings & Genes, 4: Loadings & Genes for Factor \code{factor.plot}, 5: Compound (column) Profiles}

\item{factor.plot}{Which factor (only one) should be investigated? If \code{NULL}, you can choose a factor of interest interactively from reference loadings plot.}

\item{column.interest}{Numeric vector of indices of query columns which should be in the compound profiles plot (\code{which=5}). If \code{NULL}, you can interactively select genes on the Gene Scores plot.}

\item{color.columns}{Vector of colors for the reference and query columns (compounds). If \code{NULL}, blue will be used for reference and black for query. Use this option to highlight reference columns and query columns of interest.}

\item{row.interest}{Single numeric vector or list of maximum 5 numeric vectors. This highlights gene of interest in gene scores plot (\code{which=4}) up to 5 different colors.}

\item{gene.thresP}{Threshold for genes with a high score (\code{which=4}).}

\item{gene.thresN}{Threshold for genes with a low score (\code{which=4}).}

\item{thresP.col}{Color of genes above \code{gene.thresP}.}

\item{thresN.col}{Color of genes below \code{gene.thresN}.}

\item{legend.names}{Option to draw a legend of colored columns in Compound Loadings plot (\code{which=4}). If \code{NULL}, only "References" will be in the legend.}

\item{legend.cols}{Colors to be used in legend in \code{which=4}. If \code{NULL}, only blue for "References is used".}

\item{result.available}{You can a previously returned object by \code{CSanalysis} in order to only draw graphs, not recompute the scores.}

\item{basefilename}{Basename of the graphs if saved in pdf files}

\item{plot.type}{How should the plots be outputted? \code{"pdf"} to save them in pdf files, \code{device} to draw them in a graphics device (default), \code{sweave} to use them in a sweave or knitr file.}
}
\value{
An object of the S4 Class \code{CSresult}.
}
\description{
Doing interactive CS analysis with MFA (Multiple Factor Analysis). Should use multiple references for this analysis.
}

