% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_structure.R
\name{dataSIM}
\alias{dataSIM}
\title{Simulated Microarray Data}
\format{A matrix with 1000 rows and 341 columns.}
\description{
A matrix containing some simulated example microarray data. 
The first 6 columns of this matrix make up the query matrix part.
}
