% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/method_structure.R
\docType{methods}
\name{CSanalysis,matrix,matrix,CSfabia-method}
\alias{CSanalysis,matrix,matrix,CSfabia-method}
\title{"CSfabia"}
\usage{
\S4method{CSanalysis}{matrix,matrix,CSfabia}(refMat, querMat,
  type = "CSfabia", p = 13, alpha = 0.01, cyc = 500, spl = 0,
  spz = 0.5, non_negative = 0, random = 1, center = 2, norm = 1,
  scale = 0, lap = 1, nL = 0, lL = 0, bL = 0, which = c(2, 4, 5, 7),
  BC.plot = NULL, column.interest = NULL, color.columns = NULL,
  row.interest = NULL, gene.thresP = 1, gene.thresN = -1,
  thresP.col = "blue", thresN.col = "red", legend.names = NULL,
  legend.cols = NULL, CSrank.refplot = FALSE, result.available = NULL,
  basefilename = "analyseFABIA", plot.type = "device")
}
\arguments{
\item{refMat}{Reference matrix (Rows = genes and columns = compounds)}

\item{querMat}{Query matrix}

\item{type}{\code{"CSfabia"}}

\item{p}{\emph{Fabia Parameter:} number of hidden factors = number of biclusters; default = 13}

\item{alpha}{\emph{Fabia Parameter:} sparseness loadings (0 - 1.0); default = 0.01}

\item{cyc}{\emph{Fabia Parameter:} number of iterations; default = 500}

\item{spl}{\emph{Fabia Parameter:} sparseness prior loadings (0 - 2.0); default = 0 (Laplace)}

\item{spz}{\emph{Fabia Parameter:} sparseness factors (0.5 - 2.0); default = 0.5 (Laplace)}

\item{non_negative}{\emph{Fabia Parameter:} Non-negative factors and loadings if non_negative > 0; default = 0}

\item{random}{\emph{Fabia Parameter:} <=0: by SVD, >0: random initialization of loadings in [-random,random]; default = 1.0}

\item{center}{\emph{Fabia Parameter:} data centering: 1 (mean), 2 (median), > 2 (mode), 0 (no); default = 2}

\item{norm}{\emph{Fabia Parameter:} data normalization: 1 (0.75-0.25 quantile), >1 (var=1), 0 (no); default = 1}

\item{scale}{\emph{Fabia Parameter:} loading vectors are scaled in each iteration to the given variance. 0.0 indicates non scaling; default = 0.0}

\item{lap}{\emph{Fabia Parameter:} minimal value of the variational parameter; default = 1.0}

\item{nL}{\emph{Fabia Parameter:} maximal number of biclusters at which a row element can participate; default = 0 (no limit)}

\item{lL}{\emph{Fabia Parameter:} maximal number of row elements per bicluster; default = 0 (no limit)}

\item{bL}{\emph{Fabia Parameter:} cycle at which the nL or lL maximum starts; default = 0 (start at the beginning)}

\item{which}{Choose one or more plots to draw. 1: Information Content of Biclusters, 2: Loadings for reference compounds, 3: Bicluster \code{BC1.plot} VS Bicluster \code{BC2.plot} : Loadings & Genes, 4: Gene Scores for Bicluster \code{BC.Plot}, 5: Loadings for Factor \code{BC.plot}, 6: Column (compound) profiles, 7: CS Rank Scores for Factor \code{BC.plot}}

\item{BC.plot}{Which biclusters should be investigated? Can be a vector of multiple (e.g. \code{c(1,3,5)}). If \code{NULL}, you can choose biclusters of interest interactively from reference loadings plot.}

\item{column.interest}{Numeric vector of indices of query columns which should be in the compound profiles plot (\code{which=6}). If \code{NULL}, you can interactively select genes on the Gene Scores plot.}

\item{color.columns}{Vector of colors for the reference and query columns (compounds). If \code{NULL}, blue will be used for reference and black for query. Use this option to highlight reference columns and query columns of interest.}

\item{row.interest}{Single numeric vector or list of maximum 5 numeric vectors. This highlights gene of interest in gene scores plot (\code{which=4}) up to 5 different colors.}

\item{gene.thresP}{Threshold for genes with a high score (\code{which=4}).}

\item{gene.thresN}{Threshold for genes with a low score (\code{which=4}).}

\item{thresP.col}{Color of genes above \code{gene.thresP}.}

\item{thresN.col}{Color of genes below \code{gene.thresN}.}

\item{legend.names}{Option to draw a legend of colored columns in Compound Loadings plot (\code{which=5}). If \code{NULL}, only "References" will be in the legend.}

\item{legend.cols}{Colors to be used in legend in \code{which=5}. If \code{NULL}, only blue for "References is used".}

\item{CSrank.refplot}{Logical value deciding if the CS Rank Scores (\code{which=6}) should also be plotted per reference (instead of only the weighted mean).}

\item{result.available}{You can a previously returned object by \code{CSanalysis} in order to only draw graphs, not recompute the scores.}

\item{basefilename}{Basename of the graphs if saved in pdf files}

\item{plot.type}{How should the plots be outputted? \code{"pdf"} to save them in pdf files, \code{device} to draw them in a graphics device (default), \code{sweave} to use them in a sweave or knitr file.}
}
\value{
An object of the S4 Class \code{CSresult}.
}
\description{
Doing interactive CS analysis with FABIA (Factor Analysis for Bicluster Acquisition). One or multiple reference compounds are possible in this analysis.
}

