% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/method_structure.R
\docType{methods}
\name{CSanalysis,matrix,matrix,CSmfa-method}
\alias{CSanalysis,matrix,matrix,CSmfa-method}
\title{"CSmfa"}
\usage{
\S4method{CSanalysis}{matrix,matrix,CSmfa}(refMat, querMat, type = "CSmfa",
  ncp = 5, weight.col.mfa = NULL, row.w = NULL, which = c(2, 4, 6),
  factor.plot = NULL, column.interest = NULL, row.interest = NULL,
  profile.type = "gene", color.columns = NULL, gene.highlight = NULL,
  gene.thresP = 1, gene.thresN = -1, thresP.col = "blue",
  thresN.col = "red", legend.names = NULL, legend.cols = NULL,
  CSrank.refplot = FALSE, result.available = NULL,
  basefilename = "analyseMFA", plot.type = "device")
}
\arguments{
\item{refMat}{Reference matrix (Rows = genes and columns = compounds)}

\item{querMat}{Query matrix}

\item{type}{\code{"CSmfa"}}

\item{ncp}{\emph{MFA Parameter:} Number of dimensions kept in the results (by default 5).}

\item{weight.col.mfa}{\emph{MFA Parameter:} Vector of weights, useful for HMFA method (by default, \code{NULL} and an MFA is performed).}

\item{row.w}{\emph{MFA Parameter:} An optional row weights (by default, a vector of 1 for uniform row weights).}

\item{which}{Choose one or more plots to draw:\cr 1. Percentage of variance explained by factors\cr 2. Loadings for reference compounds\cr 3. Factor 1 VS Factor 2 : Loadings & Genes\cr 4. Loadings & Genes for Factor \code{factor.plot}\cr 5. Profile plot (see \code{profile.type})\cr 6. CS Rank Scores for Factor \code{factor.plot}}

\item{factor.plot}{Which factor (only one) should be investigated? If \code{NULL}, you can choose a factor of interest interactively from reference loadings plot.}

\item{column.interest}{Numeric vector of indices of query columns which should be in the profiles plot (\code{which=5}). If \code{NULL}, you can interactively select genes in the Compound Loadings plot (\code{which=4}).}

\item{row.interest}{Numeric vector of gene indices to be plotted in gene profiles plot (\code{which=5}, \code{profile.type="gene"}). If \code{NULL}, you can interactively select them in the gene scores plot (\code{which=4}).}

\item{profile.type}{Type of \code{which=5} plot:\cr - \code{"gene"}: Gene profiles plot of selected genes in \code{row.interest} with the reference compounds and those selected in \code{column.interest} ordered first on the x axis. The other compounds are ordered in decreasing CScore. \cr - \code{"cmpd"}: Compound profiles plot of reference and selected compounds (\code{column.interest}) and only those genes on the x-axis which beat the thresholds (\code{gene.thresP}, \code{gene.thresN})}

\item{color.columns}{Vector of colors for the reference and query columns (compounds). If \code{NULL}, blue will be used for reference and black for query. Use this option to highlight reference columns and query columns of interest.}

\item{gene.highlight}{Single numeric vector or list of maximum 5 numeric vectors. This highlights gene of interest in gene scores plot (\code{which=4}) up to 5 different colors.}

\item{gene.thresP}{Threshold for genes with a high score (\code{which=4}).}

\item{gene.thresN}{Threshold for genes with a low score (\code{which=4}).}

\item{thresP.col}{Color of genes above \code{gene.thresP}.}

\item{thresN.col}{Color of genes below \code{gene.thresN}.}

\item{legend.names}{Option to draw a legend of colored columns in Compound Loadings plot (\code{which=4}). If \code{NULL}, only "References" will be in the legend.}

\item{legend.cols}{Colors to be used in legend in \code{which=4}. If \code{NULL}, only blue for "References is used".}

\item{CSrank.refplot}{Logical value deciding if the CS Rank Scores (\code{which=6}) should also be plotted per reference (instead of only the weighted mean).}

\item{result.available}{You can a previously returned object by \code{CSanalysis} in order to only draw graphs, not recompute the scores.}

\item{basefilename}{Basename of the graphs if saved in pdf files}

\item{plot.type}{How should the plots be outputted? \code{"pdf"} to save them in pdf files, \code{device} to draw them in a graphics device (default), \code{sweave} to use them in a sweave or knitr file.}
}
\value{
An object of the S4 Class \code{CSresult}.
}
\description{
Doing interactive CS analysis with MFA (Multiple Factor Analysis). Should use multiple references for this analysis.
}

