% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/method_structure.R
\docType{methods}
\name{CSanalysis,matrix,matrix,CSpca-method}
\alias{CSanalysis,matrix,matrix,CSpca-method}
\title{"CSpca"}
\usage{
\S4method{CSanalysis}{matrix,matrix,CSpca}(refMat, querMat, type = "CSpca",
  ncp = 5, scale.unit = TRUE, row.w = NULL, col.w = NULL, which = c(2,
  4, 5, 7), factor.plot = NULL, column.interest = NULL,
  row.interest = NULL, profile.type = "gene", color.columns = NULL,
  gene.highlight = NULL, gene.thresP = 1, gene.thresN = -1,
  thresP.col = "blue", thresN.col = "red", legend.names = NULL,
  legend.cols = NULL, CSrank.refplot = FALSE, result.available = NULL,
  basefilename = "analysePCA", plot.type = "device")
}
\arguments{
\item{refMat}{Reference matrix (Rows = genes and columns = compounds)}

\item{querMat}{Query matrix}

\item{type}{\code{"CSpca"}}

\item{ncp}{\emph{PCA Parameter:} Number of dimensions kept in the results (by default 5).}

\item{scale.unit}{\emph{PCA Parameter:} A boolean, if TRUE (value set by default) then data are scaled to unit variance.}

\item{row.w}{\emph{PCA Parameter:} An optional row weights (by default, a vector of 1 for uniform row weights).}

\item{col.w}{\emph{PCA Parameter:} An optional column weights (by default, uniform column weights).}

\item{which}{Choose one or more plots to draw\cr 1. Percentage of variance explained by PC's\cr 2. Loadings for reference compounds\cr 3. PC 1 VS PC 2 : Loadings & Genes\cr 4. Genes for PC \code{factor.plot}\cr 5. Loadings for PC \code{factor.plot}\cr 6. Profile plot (see \code{profile.type})\cr 7. CS Rank Scores for Factor \code{factor.plot}}

\item{factor.plot}{Which PC (only one) should be investigated? If \code{NULL}, you can choose a PC of interest interactively from reference loadings plot.}

\item{column.interest}{Numeric vector of indices of query columns which should be in the profiles plot (\code{which=5}). If \code{NULL}, you can interactively select genes on the Compound Loadings plot (\code{which=4}).}

\item{row.interest}{Numeric vector of gene indices to be plotted in gene profiles plot (\code{which=6}, \code{profile.type="gene"}). If \code{NULL}, you can interactively select them in the gene scores plot (\code{which=4}).}

\item{profile.type}{Type of \code{which=6} plot:\cr - \code{"gene"}: Gene profiles plot of selected genes in \code{row.interest} with the reference compounds and those selected in \code{column.interest} ordered first on the x axis. The other compounds are ordered in decreasing CScore. \cr - \code{"cmpd"}: Compound profiles plot of reference and selected compounds (\code{column.interest}) and only those genes on the x-axis which beat the thresholds (\code{gene.thresP}, \code{gene.thresN})}

\item{color.columns}{Vector of colors for the reference and query columns (compounds). If \code{NULL}, blue will be used for reference and black for query. Use this option to highlight reference columns and query columns of interest.}

\item{gene.highlight}{Single numeric vector or list of maximum 5 numeric vectors. This highlights gene of interest in gene scores plot (\code{which=4}) up to 5 different colors.}

\item{gene.thresP}{Threshold for genes with a high score (\code{which=4}).}

\item{gene.thresN}{Threshold for genes with a low score (\code{which=4}).}

\item{thresP.col}{Color of genes above \code{gene.thresP}.}

\item{thresN.col}{Color of genes below \code{gene.thresN}.}

\item{legend.names}{Option to draw a legend of colored columns in Compound Loadings plot (\code{which=5}). If \code{NULL}, only "References" will be in the legend.}

\item{legend.cols}{Colors to be used in legend in \code{which=5}. If \code{NULL}, only blue for "References is used".}

\item{CSrank.refplot}{Logical value deciding if the CS Rank Scores (\code{which=7}) should also be plotted per reference (instead of only the weighted mean).}

\item{result.available}{You can a previously returned object by \code{CSanalysis} in order to only draw graphs, not recompute the scores.}

\item{basefilename}{Basename of the graphs if saved in pdf files}

\item{plot.type}{How should the plots be outputted? \code{"pdf"} to save them in pdf files, \code{device} to draw them in a graphics device (default), \code{sweave} to use them in a sweave or knitr file.}
}
\value{
An object of the S4 Class \code{CSresult}.
}
\description{
Doing interactive CS analysis with PCA (Principal Component Analysis). This analysis is meant for 1 reference signature.
}

