% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/method_structure.R
\docType{class}
\name{CSresult-class}
\alias{CSresult-class}
\title{An S4 class in which the results of the Connectivity Scores by Factor Analysis are stored.}
\description{
An S4 class in which the results of the Connectivity Scores by Factor Analysis are stored.
}
\section{Slots}{

\describe{
\item{\code{type}}{A character string containing the analysis type.}

\item{\code{CS}}{List containing the connectivity scores for the query (and reference loadings). If permutation was applied, will also contain p-values.}

\item{\code{GS}}{Dataframe containing the gene scores.}

\item{\code{object}}{Object containing the FA or Zhang result.}

\item{\code{permutation.object}}{Contains permuted data, CS for permuted data and a dataframe with the p-values (only for MFA and Zhang).}

\item{\code{call}}{List object containing the original call of \code{CSanalysis} as well as the parameters for the chosen method.}
}}

